New SeedData with read products

+++ EcommerceAPI.Business/src/DTO/ProductDTO.cs
@@ -7,7 +7,7 @@ namespace EcommerceAPI.Business.src.DTO
         public Guid Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         // public Guid CategoryId { get; set; }
         public CategoryReadDTO Category { get; set; }
         public ICollection<Variant> Variants { get; set; }
@@ -23,7 +23,7 @@ namespace EcommerceAPI.Business.src.DTO
         public Guid Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public CategoryReadDTO Category { get; set; }
         public ICollection<Variant> Variants { get; set; }
         public ICollection<Image> Images { get; set; }
@@ -33,7 +33,7 @@ namespace EcommerceAPI.Business.src.DTO
     {
         public string Title { get; set; }
         public string Description { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public Guid CategoryId { get; set; }
         public ICollection<VariantCreateDTO> Variants { get; set; }
         public ICollection<Image> Images { get; set; }
@@ -46,7 +46,7 @@ namespace EcommerceAPI.Business.src.DTO
     {
         public string? Title { get; set; }
         public string? Description { get; set; }
-        public decimal? Price { get; set; }
+        public double? Price { get; set; }
 
         // Implement later: maybe also change the categoryId?
         // public ICollection<Variant> Variants { get; set; }
